Achievements
Removed the code related to account bound achievements. It wasn't working as I intended it to and it's not worth the effort to fix.

I'm reworking most of my modules since I'll be stepping back into C++ modules rather than Eluna LUA scripts. There are just too many features I want that takes too long to implement into Eluna, sadly because I have to wait for changes to be approved. I just don't have the energy to deal with all that.