Heirlooms
Added support for binding heirlooms to the account. It's very experimental and needs some serious testing, which is why it's disabled by default.

In order to access your bound heirlooms, you have to spawn the included NPC and talk to him. He will send you a vendor list of all your bound heirlooms. If you don't have any, it will not open that window and tell you that you don't have any.