Linked accounts
Added the ability to link accounts together so they share the same mounts, companions and heirlooms.

I made this to simulate the battle.net account bound feature.

There is currently no way to link accounts via in-game accounts but that might be added in the future if the feature works just as it should. The only way to link accounts is by adding them to the account_bound_linked_accounts table. The account_id column is an id of an account and the linked_account_id column is the account id that's linked to that account.

WARNING: THIS IS EXTREMELY EXPERIMENTAL AND SHOULD BE USED WITH CAUTION!