Rework
Reworked the module to support adding it to already active realms. Mounts and companions are now saved whenever the character is saved.
With that idea I decided to rewrite part of it to merge the mounts (and companions) into a single query instead of one query per spell to minimize the demand on the database. It needs a lot of testing, of course.